pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "the-map"
version = "1.0.0"
description = "Empirical EHR conformance — per-vendor, per-element deviations from US Core 6.1, with citations and verified dates."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Mock Health, Inc.", email = "hello@mock.health" },
]
keywords = [
    "fhir", "us-core", "uscdi", "epic", "cerner",
    "meditech", "interoperability", "smart-on-fhir", "hti-1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "requests>=2.32",
    "pyjwt>=2.10",
    "cryptography>=44",
    "python-dotenv>=1.0",
    "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-requests",
    "types-jsonschema",
]
llm = [
    # Optional: only needed for tools.llm_disambiguate (LLM-assisted hospital
    # resolution on top of the deterministic POS join). The deterministic
    # pipeline (build_pos_hospital_index, resolve_endpoints_to_pos) does not
    # require an API key.
    "anthropic>=0.45",
]

[project.urls]
Homepage = "https://mock.health/map"
Repository = "https://github.com/mock-health/the-map"
Changelog = "https://github.com/mock-health/the-map/blob/main/CHANGELOG.md"
Issues = "https://github.com/mock-health/the-map/issues"

[tool.setuptools]
# This is a CLI tooling repo; we ship `tools/` as importable but don't
# package the data dirs (ehrs/, us-core/, uscdi/, schema/, reports/, tests/)
# into a wheel. Downstream consumers grab those via `make export`,
# not via pip install. See EXPORT.md.
packages = ["tools", "tools.auth_flows", "tools.conformance"]

[tool.setuptools.package-data]
"tools" = ["hapi/docker-compose.example.yml"]

# ============================================================================
# ruff — fast Python linter + formatter
# ============================================================================
[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = [".venv", "tests/golden", "us-core/ig-package", "reports"]

[tool.ruff.lint]
# Stay focused on real bugs and style consistency. Avoid noisy categories
# (D for docstrings, ANN for function-annotations) that punish CLI tooling.
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes (unused imports, undefined names)
    "I",   # isort (import ordering)
    "B",   # flake8-bugbear (likely bugs)
    "UP",  # pyupgrade (modern syntax)
    "RUF", # ruff-specific
]
ignore = [
    "E501",  # line too long — many overlay literals are long by necessity
    "E741",  # ambiguous variable name (l, I, O) — used in legitimate contexts (e.g., `l` for line)
    "B007",  # unused loop control variable — `for k, v in dict.items()` skipping k is fine
    "B008",  # function call in default arg — argparse uses this pattern
    "RUF001", # ambiguous unicode in string — × and similar are intentional in FHIR notation
    "RUF002", # ambiguous unicode in docstring — same as above
    "RUF005", # collection-literal-concatenation — readability is debatable here
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]  # asserts in tests are fine
"tools/synthesize.py" = ["F841"]  # unused variables in synthesis pipeline (intentional sentinel patterns)
"tools/measure_phase_b.py" = ["F841", "RUF059"]  # unused-unpack in tuple destructuring

# ============================================================================
# mypy — type checking. Permissive for v1.0; tighten over time.
# ============================================================================
[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
explicit_package_bases = true
namespace_packages = true
mypy_path = "."
exclude = ["\\.venv", "tests/golden", "us-core/ig-package", "reports"]

# CLI tooling. We don't gate launch on type-correctness, but we do flag
# obvious mistakes. Tighten module-by-module post-launch.
[[tool.mypy.overrides]]
module = "tools.*"
disallow_untyped_defs = false
check_untyped_defs = true

# ============================================================================
# pytest — test runner config
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Don't recurse into golden fixture trees — they're data, not tests.
norecursedirs = ["tests/golden", "tests/fixtures"]
addopts = [
    "-ra",                    # show short test summary for all but pass
    "--strict-markers",
    "--tb=short",
]
filterwarnings = [
    "error",                  # promote warnings to errors in tests
    "ignore::DeprecationWarning:dotenv",  # python-dotenv deprecation noise
]
